backend/requirements.txt
# backend/requirements.txt

fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
pydantic[email]
email-validator
google-generativeai
ultralytics
geopy
geoalchemy2